+++ src/c/kamada_kawai/seq.c
@@ -4,100 +4,100 @@
 
 #include "./kamada_kawai.h"
 
-double derivaitve_x_m(double **coords, double **l_ij, double **k_ij, int index, int n)
+float derivaitve_x_m(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double sum = 0;
+    float sum = 0;
 
     for (int i = 0; i < n; i++)
     {
         if (i == index)
             continue;
 
-        double dist_x = coords[index][0] - coords[i][0];
-        double dist_y = coords[index][1] - coords[i][1];
+        float dist_x = coords[index][0] - coords[i][0];
+        float dist_y = coords[index][1] - coords[i][1];
 
-        sum += k_ij[index][i] * (dist_x - ((l_ij[index][i] * dist_x) / ((double)pow((double)pow(dist_x, 2) + (double)pow(dist_y, 2), (double)1 / 2))));
+        sum += k_ij[index][i] * (dist_x - ((l_ij[index][i] * dist_x) / ((float)pow((float)pow(dist_x, 2) + (float)pow(dist_y, 2), (float)1 / 2))));
     }
 
     return sum;
 }
 
-double derivaitve_y_m(double **coords, double **l_ij, double **k_ij, int index, int n)
+float derivaitve_y_m(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double sum = 0;
+    float sum = 0;
 
     for (int i = 0; i < n; i++)
     {
         if (i == index)
             continue;
 
-        double dist_x = coords[index][0] - coords[i][0];
-        double dist_y = coords[index][1] - coords[i][1];
+        float dist_x = coords[index][0] - coords[i][0];
+        float dist_y = coords[index][1] - coords[i][1];
 
-        sum += k_ij[index][i] * (dist_y - ((l_ij[index][i] * dist_y) / ((double)pow((double)pow(dist_x, 2) + (double)pow(dist_y, 2), (double)1 / 2))));
+        sum += k_ij[index][i] * (dist_y - ((l_ij[index][i] * dist_y) / ((float)pow((float)pow(dist_x, 2) + (float)pow(dist_y, 2), (float)1 / 2))));
     }
 
     return sum;
 }
 
-double derivaitve_xx_m(double **coords, double **l_ij, double **k_ij, int index, int n)
+float derivaitve_xx_m(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double sum = 0;
+    float sum = 0;
 
     for (int i = 0; i < n; i++)
     {
         if (i == index)
             continue;
 
-        double dist_x = coords[index][0] - coords[i][0];
-        double dist_y = coords[index][1] - coords[i][1];
+        float dist_x = coords[index][0] - coords[i][0];
+        float dist_y = coords[index][1] - coords[i][1];
 
-        sum += k_ij[index][i] * (1 - ((l_ij[index][i] * (double)pow(dist_y, 2)) / ((double)pow((double)pow(dist_x, 2) + (double)pow(dist_y, 2), (double)3 / 2))));
+        sum += k_ij[index][i] * (1 - ((l_ij[index][i] * (float)pow(dist_y, 2)) / ((float)pow((float)pow(dist_x, 2) + (float)pow(dist_y, 2), (float)3 / 2))));
     }
 
     return sum;
 }
 
-double derivaitve_yy_m(double **coords, double **l_ij, double **k_ij, int index, int n)
+float derivaitve_yy_m(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double sum = 0;
+    float sum = 0;
 
     for (int i = 0; i < n; i++)
     {
         if (i == index)
             continue;
 
-        double dist_x = coords[index][0] - coords[i][0];
-        double dist_y = coords[index][1] - coords[i][1];
+        float dist_x = coords[index][0] - coords[i][0];
+        float dist_y = coords[index][1] - coords[i][1];
 
-        sum += k_ij[index][i] * (1 - ((l_ij[index][i] * (double)pow(dist_x, 2)) / ((double)pow((double)pow(dist_x, 2) + (double)pow(dist_y, 2), (double)3 / 2))));
+        sum += k_ij[index][i] * (1 - ((l_ij[index][i] * (float)pow(dist_x, 2)) / ((float)pow((float)pow(dist_x, 2) + (float)pow(dist_y, 2), (float)3 / 2))));
     }
 
     return sum;
 }
 
-double derivaitve_xy_m(double **coords, double **l_ij, double **k_ij, int index, int n)
+float derivaitve_xy_m(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double sum = 0;
+    float sum = 0;
 
     for (int i = 0; i < n; i++)
     {
         if (i == index)
             continue;
 
-        double dist_x = coords[index][0] - coords[i][0];
-        double dist_y = coords[index][1] - coords[i][1];
+        float dist_x = coords[index][0] - coords[i][0];
+        float dist_y = coords[index][1] - coords[i][1];
 
-        sum += k_ij[index][i] * ((l_ij[index][i] * dist_x * dist_y) / ((double)pow((double)pow(dist_x, 2) + (double)pow(dist_y, 2), (double)3 / 2)));
+        sum += k_ij[index][i] * ((l_ij[index][i] * dist_x * dist_y) / ((float)pow((float)pow(dist_x, 2) + (float)pow(dist_y, 2), (float)3 / 2)));
     }
 
     return sum;
 }
 
-int get_max_delta_m_index(double *deltas, int n, double epsilon)
+int get_max_delta_m_index(float *deltas, int n, float epsilon)
 {
     int max_index = -1;
-    double max = -1;
+    float max = -1;
 
     for (int i = 0; i < n; i++)
     {
@@ -112,22 +112,22 @@ int get_max_delta_m_index(double *deltas, int n, double epsilon)
     return max_index;
 }
 
-double delta_m(double derivaitve_x, double derivaitve_y)
+float delta_m(float derivaitve_x, float derivaitve_y)
 {
-    return sqrt((double)pow(derivaitve_x, 2) + (double)pow(derivaitve_y, 2));
+    return sqrt((float)pow(derivaitve_x, 2) + (float)pow(derivaitve_y, 2));
 }
 
-double calculate_delta(double **coords, double **l_ij, double **k_ij, int index, int n)
+float calculate_delta(float **coords, float **l_ij, float **k_ij, int index, int n)
 {
-    double derivaitve_x = derivaitve_x_m(coords, l_ij, k_ij, index, n);
-    double derivaitve_y = derivaitve_y_m(coords, l_ij, k_ij, index, n);
+    float derivaitve_x = derivaitve_x_m(coords, l_ij, k_ij, index, n);
+    float derivaitve_y = derivaitve_y_m(coords, l_ij, k_ij, index, n);
 
     return delta_m(derivaitve_x, derivaitve_y);
 }
 
-double *calculate_delatas(double **coords, double **l_ij, double **k_ij, int n)
+float *calculate_delatas(float **coords, float **l_ij, float **k_ij, int n)
 {
-    double *deltas = (double *)malloc(n * sizeof(double));
+    float *deltas = (float *)malloc(n * sizeof(float));
 
     for (int i = 0; i < n; i++)
     {
@@ -137,21 +137,21 @@ double *calculate_delatas(double **coords, double **l_ij, double **k_ij, int n)
     return deltas;
 }
 
-double calculate_delta_y(double derivaitve_x_m, double derivaitve_y_m, double derivaitve_xx_m, double derivaitve_yy_m, double derivaitve_xy_m)
+float calculate_delta_y(float derivaitve_x_m, float derivaitve_y_m, float derivaitve_xx_m, float derivaitve_yy_m, float derivaitve_xy_m)
 {
     return (-(derivaitve_xy_m * derivaitve_x_m) + (derivaitve_xx_m * derivaitve_y_m)) / (-(derivaitve_xx_m * derivaitve_yy_m) + (derivaitve_xy_m * derivaitve_xy_m));
 }
 
-double calculate_delta_x(double derivaitve_x_m, double derivaitve_y_m, double derivaitve_xx_m, double derivaitve_yy_m, double derivaitve_xy_m, double delta_y)
+float calculate_delta_x(float derivaitve_x_m, float derivaitve_y_m, float derivaitve_xx_m, float derivaitve_yy_m, float derivaitve_xy_m, float delta_y)
 {
     return (-(derivaitve_y_m) - (derivaitve_yy_m * delta_y)) / derivaitve_xy_m;
 }
 
-void copyCoords(double **coords, double **vertices, int n)
+void copyCoords(float **coords, float **vertices, int n)
 {
     for (int i = 0; i < n; i++)
     {
-        vertices[i] = (double *)malloc(2 * sizeof(double));
+        vertices[i] = (float *)malloc(2 * sizeof(float));
         vertices[i][0] = coords[i][0];
         vertices[i][1] = coords[i][1];
     }
@@ -160,14 +160,14 @@ void copyCoords(double **coords, double **vertices, int n)
 Vertices *seq(KamadaKawai *kk)
 {
     Vertices *vertices = (Vertices *)malloc(sizeof(Vertices));
-    vertices->coordinates = (double **)malloc(kk->n * sizeof(double *));
+    vertices->coordinates = (float **)malloc(kk->n * sizeof(float *));
     copyCoords(kk->coordinates, vertices->coordinates, kk->n);
 
     Vertices *vertices_head = vertices;
     
-    double epsilon = kk->epsilon;
+    float epsilon = kk->epsilon;
 
-    double *deltas = calculate_delatas(kk->coordinates, kk->l_ij, kk->k_ij, kk->n);
+    float *deltas = calculate_delatas(kk->coordinates, kk->l_ij, kk->k_ij, kk->n);
     // print deltas here
     // for (int i = 0; i < kk->n; i++)
     // {
@@ -179,20 +179,20 @@ Vertices *seq(KamadaKawai *kk)
     {
         while (deltas[max_delta_m_index] > epsilon)
         {
-            double d_x_m = derivaitve_x_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
-            double d_y_m = derivaitve_y_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
-            double d_xx_m = derivaitve_xx_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
-            double d_yy_m = derivaitve_yy_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
-            double d_xy_m = derivaitve_xy_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
+            float d_x_m = derivaitve_x_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
+            float d_y_m = derivaitve_y_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
+            float d_xx_m = derivaitve_xx_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
+            float d_yy_m = derivaitve_yy_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
+            float d_xy_m = derivaitve_xy_m(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
 
-            double delta_y = calculate_delta_y(
+            float delta_y = calculate_delta_y(
                 d_x_m,
                 d_y_m,
                 d_xx_m,
                 d_yy_m,
                 d_xy_m);
 
-            double delta_x = calculate_delta_x(
+            float delta_x = calculate_delta_x(
                 d_x_m,
                 d_y_m,
                 d_xx_m,
@@ -206,7 +206,7 @@ Vertices *seq(KamadaKawai *kk)
             deltas[max_delta_m_index] = calculate_delta(kk->coordinates, kk->l_ij, kk->k_ij, max_delta_m_index, kk->n);
         }
         vertices->next = (Vertices *)malloc(sizeof(Vertices));
-        vertices->next->coordinates = (double **)malloc(kk->n * sizeof(double *));
+        vertices->next->coordinates = (float **)malloc(kk->n * sizeof(float *));
         copyCoords(kk->coordinates, vertices->next->coordinates, kk->n);
 
         vertices = vertices->next;
