+++ src/c/lib/types.h
@@ -11,24 +11,24 @@ typedef struct dimensions
 typedef struct KamadaKawai
 {
     int n;                // number of vertices
-    double **coordinates; // coordinates of the vertices
+    float **coordinates; // coordinates of the vertices
 
     int m;       // number of edges
     int **edges; // edges between the vertices
 
     int **d_ij;
-    double **l_ij;
-    double **k_ij;
+    float **l_ij;
+    float **k_ij;
 
-    double K;       // constant k
-    double epsilon; // epsilon
-    double display; // length of a side of a display square area
+    float K;       // constant k
+    float epsilon; // epsilon
+    float display; // length of a side of a display square area
 } KamadaKawai;
 
 typedef struct Vertices
 {
     struct Vertices *next;
-    double **coordinates;
+    float **coordinates;
 } Vertices;
 
 #endif
